modif facturation #1

+++ src/Controller/DossierMainOeuvreController.php
@@ -22,6 +22,17 @@ class DossierMainOeuvreController extends MyAdminController
 		$response = parent::newAction();
         $entity = $this->request->attributes->get('easyadmin')['item'];
         if ($response instanceof RedirectResponse) {
+            $devis_array = $entity->getDossier()->getDevis();
+            if(sizeof($devis_array) == 1){
+                $devis = $devis_array[0];
+                if($devis->getStatut() == 0 && $devis->getNewDevis() == 1)
+                {
+                    $entity->addNewDevis($devis);
+                    $devis->addNewDossierMainOeuvre($entity);
+                    $this->em->flush();
+                }
+            }
+
             return $this->redirectToRoute('admin', ['entity' => 'Dossier', 'action' => 'show', 'id' => $entity->getDossier()->getId()]);
         }
 
@@ -41,14 +52,40 @@ class DossierMainOeuvreController extends MyAdminController
 
     protected function deleteAction()
     {
-        $response = parent::deleteAction();
-
         $entity = $this->request->attributes->get('easyadmin')['item'];
-        if ($response instanceof RedirectResponse) {
-            return $this->redirectToRoute('admin', ['entity' => 'Dossier', 'action' => 'show', 'id' => $entity->getDossier()->getId()]);
+        $devis_array = $entity->getDossier()->getDevis();
+        $allow_delete = true;
+        if(sizeof($devis_array) > 0){
+            foreach($devis_array as $devis){
+                if($devis->getDossierMainOeuvres()->contains($entity) && $devis->getStatut() == 1 && $devis->getNewDevis() == 1){
+                    $allow_delete = false;
+                }
+            }
         }
 
-        return $response;
+        if($allow_delete)
+        {
+            $response = parent::deleteAction();
+            
+            if ($response instanceof RedirectResponse) {               
+                if(sizeof($devis_array) == 1){
+                    $devis = $devis_array[0];
+                    if($devis->getStatut() == 0 && $devis->getNewDevis() == 1)
+                    {
+                        $devis->removeDossierMainOeuvre($entity);
+                        $this->em->flush();
+                    }
+
+                }
+
+                return $this->redirectToRoute('admin', ['entity' => 'Dossier', 'action' => 'show', 'id' => $entity->getDossier()->getId()]);
+            }
+
+            return $response;
+        }else{
+            $this->addFlash("error", "Cette main d'oeuvre ne peux pas être supprimer car elle a déjà été facturé");
+            return $this->redirectToRoute('admin', ['entity' => 'Dossier', 'action' => 'show', 'id' => $entity->getDossier()->getId()]);
+        }
     }
 
     protected function listAction()
